fixed CSS issues and updated footer
code was added to style the <p> tags in the body of the HTML files as
well as to style the 'return to home' buttons on the email page. The
black stroke in the navbar was deleted and a photo was added to
'about-us'. The email pages were also updated in both HTML and CSS.